requirements.txt
Flask
requests
Pillow
numpy
colour-science